// enums/firebase-errors-messages.ts
export enum FirebaseErrorMessages {
  WEAK_PASSWORD = "auth/weak-password",
  EMAIL_ALREADY_IN_USE = "auth/email-already-in-use",
  INVALID_LOGIN_CREDENTIALS = "auth/invalid-login-credentials",
  INVALID_CREDENTAILS = "auth/invalid-credential",
  WRONG_PASSWORD = "auth/wrong-password",
  USER_NOT_FOUND = "auth/user-not-found",
}